refactor: parseHtml will be used in multiple places
